// stay-in-sync-configurator-ui/src/app/features/source-system/generated/model/apiEndpointQueryParamType.ts
/**
 * core-management API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const ApiEndpointQueryParamType = {
    Path: 'PATH',
    Query: 'QUERY'
} as const;
export type ApiEndpointQueryParamType = typeof ApiEndpointQueryParamType[keyof typeof ApiEndpointQueryParamType];